+++ out/extension.js
@@ -40,6 +40,9 @@ function activate(context) {
     vscode.commands.registerCommand('aws-access-vscode-extension.TestAwsConnectivity', () => {
         StatusBar.StatusBarItem.Current.TestAwsConnectivity();
     });
+    vscode.commands.registerCommand('aws-access-vscode-extension.CopyCredentialsToDefaultProfile', () => {
+        StatusBar.StatusBarItem.Current.CopyCredentialsToDefaultProfile();
+    });
     vscode.window.onDidCloseTerminal((terminal) => {
         StatusBar.StatusBarItem.Current.onDidCloseTerminal(terminal);
     });